move prompts and classes to separate files
