+++ effect_creator.py
@@ -1,27 +1,13 @@
-import json
-from pydantic import BaseModel
-from typing import Literal, Optional, List
 from llm_tools import call_llm
+from llm_prompts import MAGIC_PROMPT_TEMPLATE
 
-MAGIC_PROMPT_TEMPLATE = """
-You are the governing force of magic in a fantasy universe. Given a description of a set of circumstances that a group of mortals in your universe have constructed, determine what type of magical effect will occur.
-If there is no effect respond with: there is no magical effect.
-Description:
-"{description}"
-
-
-"""
-
-
-
-def create_effect(description, system="D&D 5e"):
+def create_effect(description: str):
 
     prompt = MAGIC_PROMPT_TEMPLATE.format(description=description)
     print("Sending to LLM...")
 
     try:
         response = call_llm(prompt)
-        print(response)
         print(f"Effect: {response}")
     except Exception as e:
         print(f"Error: {e}")
